feat(rename): rename the project to babel-plugin-macros
Closes #41

BREAKING CHANGE: upgrades to babel v7 and renames to babel-plugin-macros
