+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ Thanks for being willing to contribute!
 > pull requests from branches on your fork. To do this, run:
 >
 > ```
-> git remote add upstream https://github.com/kentcdodds/babel-macros.git
+> git remote add upstream https://github.com/kentcdodds/babel-plugin-macros.git
 > git fetch upstream
 > git branch --set-upstream-to=upstream/master master
 > ```
@@ -69,4 +69,4 @@ requests! Thanks!
 
 [egghead]: https://egghead.io/series/how-to-contribute-to-an-open-source-project-on-github
 [all-contributors]: https://github.com/kentcdodds/all-contributors
-[issues]: https://github.com/kentcdodds/babel-macros/issues
+[issues]: https://github.com/kentcdodds/babel-plugin-macros/issues
