+++ other/mock-modules/babel-plugin-macros-test-error-thrower.macro/index.js
