+++ other/mock-modules/babel-plugin-macros-test-fake/macro.js
