+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "babel-macros",
+  "name": "babel-plugin-macros",
   "version": "0.0.0-semantically-released",
   "description": "Enables zero-config, importable babel plugins",
   "main": "dist/index.js",
@@ -27,26 +27,26 @@
     "cosmiconfig": "3.1.0"
   },
   "devDependencies": {
-    "ast-pretty-print": "2.0.1",
-    "babel-core": "^6.25.0",
-    "babel-plugin-tester": "5.0.0",
-    "babylon": "6.18.0",
-    "cpy": "6.0.0",
-    "kcd-scripts": "0.30.4"
+    "ast-pretty-print": "^2.0.1",
+    "babel-core": "7.0.0-beta.3",
+    "babel-plugin-tester": "^5.0.0",
+    "babylon": "7.0.0-beta.34",
+    "cpy": "^6.0.0",
+    "kcd-scripts": "^0.30.4"
   },
   "eslintConfig": {
     "extends": "./node_modules/kcd-scripts/eslint.js"
   },
   "eslintIgnore": ["node_modules", "coverage", "dist"],
   "babel": {
-    "presets": ["./.babelrc.js"]
+    "presets": ["./other/babel-config.js"]
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/kentcdodds/babel-macros.git"
+    "url": "https://github.com/kentcdodds/babel-plugin-macros.git"
   },
   "bugs": {
-    "url": "https://github.com/kentcdodds/babel-macros/issues"
+    "url": "https://github.com/kentcdodds/babel-plugin-macros/issues"
   },
-  "homepage": "https://github.com/kentcdodds/babel-macros#readme"
+  "homepage": "https://github.com/kentcdodds/babel-plugin-macros#readme"
 }
