+++ src/__tests__/fixtures/config/babel-plugin-macros.config.js
